Add an extra (void *) cast.  The struct if_msghdr has an 8 byte alignment requirement, while rt_msghdr has a 4 byte alignment requirement.  The root cause is that if_msghdr has an struct if_data which has an 8-byte alignment requirement due to a time_t that's embedded in it.  On MIPS, time_t is a 64-bit number, so must be 64-bit aligned.
Since we don't access ifm_data.ifi_epoch, a simple cast is all that's
necessary here.  It is likely the case that ifi_epoch should *NOT* be
a time_t because it is an uptime (time delta) an not an absolute time
since 1970.  u_long is likely sufficient there since that gives an
uptime of 136 years will suffice for the foreseeable future.


git-svn-id: http://svn.freebsd.org/base/head@188760 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f