Add the infrastructure and expected sizeof() values for each supported architecture to implement size-guards on the vimage vnet_* structures.
As CTASSERT_EQUAL() needs special compile time options we back it
by CTASSERT() in the default case. Unfortunately CTASSERT() triggers
first, thus add an option to allow compilation with CTASSERT_EQUAL() only.

See the comments how to get new values if you trigger the assert
and what to do in that case.

Reviewed by:	rwatson, zec (earlier versions)


git-svn-id: http://svn.freebsd.org/base/head@189223 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f