It turns out that the Toshiba LANCT00A PC Card is really like the Toshiba PCETC ISA card, and even has the same board type code in the card ID (0x14).  So, for this card, call ed_probe_WD80x3_generic after setting things up apropriately.  This makes the card attach and kinda work (I'm seeing panics in sbdrop).  Since history has shown that the WD80x3 probe routine is dangerous, only do it for this card.  Also, disable the memory range check to make sure it is an valid ISA memory. I think that it is bogus, but I'm not 100% sure, for these cards.
I removed probing for the WD80x3 in 2005 when I added support for the
AX88x90 and DL100xx cards since none of my cards had ever matched it
and PAO3 removed it and none of the cards in their database died.

It is possible there are other quirks about this card too, since no
other open source OS supports it, or even claims to support it.  But
it was a fun half hour hack...


git-svn-id: http://svn.freebsd.org/base/head@190483 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f