Add emulation of delay profiles, which lets you model various types of MAC overheads such as preambles, link level retransmissions and more.
Note- this commit changes the userland/kernel ABI for pipes
(but not for ordinary firewall rules) so you need to rebuild
kernel and /sbin/ipfw to use dummynet features.

Please check the manpage for details on the new feature.

The MFC would be trivial but it breaks the ABI, so it will
be postponed until after 7.2 is released.

Interested users are welcome to apply the patch manually
to their RELENG_7 tree.

Work supported by the European Commission, Projects Onelab and
Onelab2 (contract 224263).


git-svn-id: http://svn.freebsd.org/base/head@190865 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f